backend/verify_tables.py
#!/usr/bin/env python3
"""Verify database tables exist"""
import os
from dotenv import load_dotenv
from supabase import create_client

load_dotenv()

supabase_url = os.getenv("SUPABASE_URL")
service_key = os.getenv("SUPABASE_SERVICE_KEY")
supabase = create_client(supabase_url, service_key)

print("=" * 70)
print("🔍 Database Table Verification")
print("=" * 70)
print()

# List of actual tables from the migration
tables = [
    'users',
    'email_verifications',
    'phone_verifications',
    'refresh_tokens',
    'password_resets',
    'social_auth_providers',
    'auth_audit_log'
]

existing = []
missing = []

for table in tables:
    try:
        supabase.table(table).select('id').limit(0).execute()
        existing.append(table)
        print(f"✅ {table}")
    except Exception as e:
        missing.append(table)
        print(f"❌ {table} - {str(e)[:50]}")

print()
print("=" * 70)
print(f"📊 Summary: {len(existing)}/{len(tables)} tables exist")
print("=" * 70)

if len(existing) == len(tables):
    print()
    print("🎉 ALL AUTHENTICATION TABLES ARE READY!")
    print()
    print("✅ Database migration completed successfully")
    print("✅ SMS OTP system is ready for production")
    print()
    exit(0)
else:
    print()
    print(f"⚠️  {len(missing)} table(s) still missing")
    exit(1)
