backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Supabase
supabase>=2.0.0,<3.0.0
postgrest>=0.10.0

# Database
asyncpg==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
httpx<0.25.0,>=0.24.0
redis==5.0.1
slowapi==0.1.9

# WebSocket
websockets==12.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
black==23.12.1
flake8==7.0.0
loguru==0.7.2
